Разделил методы Action и CRUD на два класса
