+++ src/test/java/autotest/clients/DuckActionClients.java
@@ -14,8 +14,7 @@ import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 public class DuckActionClients extends TestNGCitrusSpringSupport {
 
     @Autowired
-    //почему не private? ведь дальше этого класса данное поле не используется?
-    protected HttpClient yellowDuckService;
+    private HttpClient yellowDuckService;
 
     public void duckFly(TestCaseRunner runner, String id) {
         runner.$(http()
@@ -53,63 +52,6 @@ public class DuckActionClients extends TestNGCitrusSpringSupport {
                 .send()
                 .get("/api/duck/action/swim")
                 .queryParam("id", id));
-    }
-
-    public void duckCreate(TestCaseRunner runner,
-                           String color,
-                           String height,
-                           String material,
-                           String sound,
-                           String wingsState) {
-        runner.$(http()
-                .client(yellowDuckService)
-                .send()
-                .post("/api/duck/create")
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body("{\n" +
-                        "  \"color\": \"" + color + "\",\n" +
-                        "  \"height\": " + height + ",\n" +
-                        "  \"material\": \"" + material + "\",\n" +
-                        "  \"sound\": \"" + sound + "\",\n" +
-                        "  \"wingsState\": \"" + wingsState + "\"\n" +
-                        "}"));
-    }
-
-    public void duckDelete(TestCaseRunner runner,
-                           String id) {
-        runner.$(http()
-                .client(yellowDuckService)
-                .send()
-                .delete("/api/duck/delete")
-                .queryParam("id", id));
-    }
-
-    public void duckGetAllIds(TestCaseRunner runner) {
-        runner.$(http()
-                .client(yellowDuckService)
-                .send()
-                .get("/api/duck/getAllIds"));
-    }
-
-    public void duckUpdate(TestCaseRunner runner,
-                           String id,
-                           String color,
-                           String height,
-                           String material,
-                           String sound,
-                           String wingsState) {
-        runner.$(http()
-                .client(yellowDuckService)
-                .send()
-                .put("/api/duck/update")
-                .message()
-                .queryParam("id", id)
-                .queryParam("color", color)
-                .queryParam("height", height)
-                .queryParam("material", material)
-                .queryParam("sound", sound)
-                .queryParam("wingsState", wingsState));
     }
 }
